+++ src/Oro/Bundle/ShippingBundle/Method/MultiShippingMethodType.php
@@ -10,7 +10,7 @@ use Oro\Bundle\ShippingBundle\Provider\MultiShippingCostProvider;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 
 /**
- * Multi Shipping method type implementation.
+ * Represents Multi Shipping method type.
  */
 class MultiShippingMethodType implements ShippingMethodTypeInterface
 {
@@ -30,26 +30,41 @@ class MultiShippingMethodType implements ShippingMethodTypeInterface
         $this->shippingCostProvider = $shippingCostProvider;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function getIdentifier(): string
     {
-        return static::IDENTIFIER;
+        return self::IDENTIFIER;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function getLabel(): string
     {
         return $this->label;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function getSortOrder(): int
     {
         return 0;
     }
 
-    public function getOptionsConfigurationFormType(): string
+    /**
+     * {@inheritDoc}
+     */
+    public function getOptionsConfigurationFormType(): ?string
     {
         return HiddenType::class;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function calculatePrice(
         ShippingContextInterface $context,
         array $methodOptions,