+++ src/Oro/Bundle/ShippingBundle/Tests/Unit/Method/Stub/TrackingAwareShippingMethodStub.php
@@ -2,79 +2,15 @@
 
 namespace Oro\Bundle\ShippingBundle\Tests\Unit\Method\Stub;
 
-use Oro\Bundle\ShippingBundle\Method\ShippingMethodInterface;
 use Oro\Bundle\ShippingBundle\Method\ShippingTrackingAwareInterface;
+use Oro\Bundle\ShippingBundle\Tests\Unit\Provider\Stub\ShippingMethodStub;
 
-class TrackingAwareShippingMethodStub implements ShippingMethodInterface, ShippingTrackingAwareInterface
+class TrackingAwareShippingMethodStub extends ShippingMethodStub implements ShippingTrackingAwareInterface
 {
     /**
      * {@inheritDoc}
      */
-    public function isEnabled()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isGrouped()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getIdentifier()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getLabel(): string
-    {
-        return '';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getTypes()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getType($identifier)
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getOptionsConfigurationFormType()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getSortOrder()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getTrackingLink($number)
+    public function getTrackingLink(string $number): ?string
     {
         return null;
     }